backend/requirements.txt
asgiref==3.8.1
boto3==1.34.140
botocore==1.34.140
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
wheel==0.43.0
